make DrawToPNG more respectable
